Create lego_wall
This problem involves an input of various "lego" blocks with different heights. The problem provides a target height, and asks to output the total amount of combinations possible using the various lego heights to for a wall of target height. Each block can be used infinite times. This solution requires dynamic programming. I have heavily commented this solution to help further understand / remember the process.